% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AD.R
\name{AD}
\alias{AD}
\title{Audio url/path to HTML Audio Tag}
\usage{
AD(audio = "clipboard", is_url = is.url(audio), copy2clip = interactive(),
  print = FALSE)
}
\arguments{
\item{audio}{A url, path or file located in the audio directory which is an
audio file.}

\item{is_url}{logical.  If \code{TRUE} \file{audio} is located in the
\file{~audio} or \file{../audio} directory of \file{REPORTS} or
\file{PRESENTATION} directories.}

\item{copy2clip}{If \code{TRUE} attempts to copy the output to the clipboard.}

\item{print}{logical. If \code{TRUE} \code{\link[base]{cat}} prints the
output to the console. If \code{FALSE} \code{\link[base]{return}}s to the
console.}
}
\value{
Returns a character vector of an HTML source tag that embeds an audio
file.
}
\description{
Converts a url/path to an HTML audio tag.  \code{AD} attempts to determine
the location of the audio file if it is not a path or url.
}
\examples{
\dontrun{
AD("foo.mp4")
}
}
\references{
\url{http://www.w3schools.com/html/html5_audio.asp}
}
\keyword{audio}

