% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BV.R
\name{BV}
\alias{BV}
\title{View Bibliography}
\usage{
BV(bib = NULL, col.width = 40)
}
\arguments{
\item{bib}{Path to bibliography.  If \code{NULL} \code{BV} attempts to locate
the .bib file (working directory first and then "main_dir/REPORTS/_.bib",
last "main_dir/PRESENTATION/_.bib" in that order).}

\item{col.width}{An integer value of the maximum width of columns.}
}
\value{
Returns a truncated view of user references (.bib file).
}
\description{
Allows a truncated view of your .bib.
}
\examples{
## BV()
}

