\name{install_pandoc}
\alias{install_pandoc}
\title{Download Pandoc}
\usage{
  install_pandoc()
}
\value{
  Installs Pandoc on your system.
}
\description{
  Windows users download Pandoc from the command line.
}
\examples{
\dontrun{
install_pandoc()
}
}
\author{
  Gergely Daroczi and Gabor Grothendieck.
}
\references{
  \url{http://stackoverflow.com/a/15072501/1000343}
  \url{http://johnmacfarlane.net/pandoc/}
}
\seealso{
  \code{\link[qdap]{dir_map}}
}
\keyword{transcript}

