% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/notes.R
\name{notes}
\alias{notes}
\title{View Notes}
\usage{
notes(col.width = 70, notes.col = FALSE)
}
\arguments{
\item{col.width}{An integer value of the maximum width of columns.}

\item{notes.col}{logical.  If \code{TRUE} the notes column will be displayed.}
}
\value{
Returns a truncated view of user notes.
}
\description{
Allows a truncated view of your notes (\file{~/ARTICLES/notes}).
}
\examples{
\dontrun{
notes()
}
}

