% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/web.R
\name{print.read.web}
\alias{print.read.web}
\title{Prints a read.web Object.}
\usage{
\method{print}{read.web}(x, ...)
}
\arguments{
\item{x}{The read.web object}

\item{\ldots}{ignored}
}
\description{
Prints a read.web object.
}

