% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/new_report.R
\name{print.reports}
\alias{print.reports}
\title{Prints a reports Object}
\usage{
\method{print}{reports}(x, ...)
}
\arguments{
\item{x}{The reports object.}

\item{\ldots}{ignored}
}
\description{
Prints a reports object.
}

