% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports-package.R
\docType{package}
\name{reports-package}
\alias{reports}
\alias{reports-package}
\title{reports: Generate Activity Reports}
\description{
Generates scientific reports in PDF format.
}
\author{
\strong{Maintainer}: Dany Mukesha \email{danymukesha@gmail.com} (\href{https://orcid.org/0009-0001-9514-751X}{ORCID})

}
\keyword{internal}
