\docType{package}
\name{reports}
\alias{package-reports}
\alias{reports}
\alias{reports-package}
\title{reports: Assists in the workflow of writing reports}
\description{
  reports is a package that assists in writing apa6 style
  reports and presentations.  The package is designed to be
  used with RStudio, MikTex/TexLive, knitr and Pandoc.  The
  user will want to download these free programs to
  maximize the effectiveness of the reports package.
}

