\docType{data}
\name{themes}
\alias{themes}
\title{themes}
\format{A character vector with 28 elements}
\description{
  a vector of available beamer themes
}
\details{
  \href{http://deic.uab.es/~iblanes/beamer_gallery/index_by_theme.html}{http://deic.uab.es/~iblanes/beamer_gallery/index_by_theme.html}
}
\references{
  \url{http://deic.uab.es/~iblanes/beamer_gallery/index_by_theme.htm}
}
\keyword{datasets}

