\name{url_dl}
\alias{url_dl}
\title{Download Instructional Documents}
\usage{
  url_dl(..., url = "http://dl.dropbox.com/u/61803503/")
}
\arguments{
  \item{\ldots}{Document names to download.}

  \item{url}{The download url.}
}
\value{
  Places a copy of the downloaded document in the users
  working directory.
}
\description{
  This function enables downloading documents for future
  instructional training.
}
\note{
  Not intended for general use.
}

