% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/custom_css.R
\name{custom_css}
\alias{css_style_add}
\alias{css_style_change}
\alias{css_styles}
\alias{custom_css}
\alias{custom_pandoc_style}
\title{Generate Custom css for RStudio + knitr}
\usage{
custom_css(rprofile = FALSE, loc = QP("REPORT"), style.css = NULL,
  source = TRUE)

css_styles()

css_style_change(style.css, cur.style.css = QP("REPORT/css/style.css"))

css_style_add(loc = QP("PRESENTATION/assets/css"))

custom_pandoc_style(rprofile = FALSE, loc = QP("REPORT"), source = TRUE)
}
\arguments{
\item{rprofile}{logical.  If \code{TRUE} the \bold{style.R} is added to the
\bold{.Rprofile} for sourcing upon startup.}

\item{loc}{Path to the report location where the custom css should be placed.
If \code{NULL} only the \bold{style.R} is created in the base directory
\bold{.Rprofile}.}

\item{style.css}{A path to a \bold{style.css} file that will be used as
the \file{~/css/style.css}.  Also may take a character string indicating one
of the reports package's built in \bold{style.css} (see
\code{\link[reports]{css_styles}}).  Default in
\code{\link[reports]{custom_css}} is a blank \bold{style.css}.}

\item{source}{logical.  If \code{TRUE} the style.R is sourced intially.}

\item{cur.style.css}{The location of the current style.css (this will be
replaced).}
}
\description{
\code{custom_css} - Generate the components necessary for a custom css for
use with RStudio and knitr.

\code{css_styles} - View the available built in style.css files.

\code{css_style_change} - Change a style.css.

\code{css_styl_add} - Add a style.css.  Defaults to
\file{~PRESENTATION/assets/css}.

\code{custom_pandoc_style} - Generate the components necessary for a custom
pandoc styling for use with RStudio and knitr.
}
\details{
\code{custom_css} - The user must add the custom contents to the
custom css located in  \file{~/css/style.css}.

\code{css_styles} - prints a list of available style.css templates
available within the reports package for use with the \code{custom_css}
argument \code{style.css}.  Feel free to submit your own to the reports
package \email{reports.rpackage@gmail.com} marked with \bold{style.css
template} in the subject line.

See \url{https://github.com/trinker/style.css_examples} for examples
utilizing the internal reports package style.css.
}
\note{
The user has to source before the custom styles will be applied. The
user may specify \code{rfprofile = TRUE} to add the \bold{style.R} to the
\bold{.Rprofile} for sourcing upon startup.  Otherwise, the user must open
the \bold{style.R} as the active tab in RStudio, check the `Source on Save`
box and the click the save icon.
}
\examples{
## custom_css(TRUE, style.css = css_styles()[1])
css_styles()
## css_style_change("rinker_vignette")
}
\references{
\url{http://www.rstudio.com/ide/docs/authoring/markdown_custom_rendering}
}

