\name{install_pandoc}
\alias{install_pandoc}
\title{Download Pandoc}
\usage{
  install_pandoc()
}
\value{
  Installs Pandoc on your system.
}
\description{
  Download Pandoc from the command line (Windows users).
}
\section{Pandoc Website}{
  \url{http://johnmacfarlane.net/pandoc/}
}
\examples{
\dontrun{
install_pandoc()
}
}
\author{
  Gergely Daroczi and Gabor Grothendieck
}
\references{
  \url{http://stackoverflow.com/a/15072501/1000343}
}

