% Generated by roxygen2 (4.0.1): do not edit by hand
\name{js_copy}
\alias{js_copy}
\title{Utility Fuction to Copy reports Package JavaScript}
\usage{
js_copy(outfile = "PRESENTATION/assets/js")
}
\arguments{
\item{outfile}{The directory to copy the JavaScript files to.}
}
\description{
Copies all the reports Package JavaScript files to designated directory.
}
\keyword{javascript}

