% Generated by roxygen2 (4.0.1): do not edit by hand
\name{LL}
\alias{LL}
\title{Format Text Lines to LaTeX List}
\usage{
LL(enumerate = TRUE, text = "clipboard", copy2clip = interactive())
}
\arguments{
\item{enumerate}{logical.  If \code{TRUE} uses the enumerate environment.  If
\code{FALSE} itemize is used instead.}

\item{text}{A character vector or text copied to the clipboard.  Default is to
read from the clipboard.}

\item{copy2clip}{logical.  If \code{TRUE} attempts to copy the output to the
clipboard.}
}
\value{
Returns a character vector with a LaTeX list formatted text.
}
\description{
Itemizes lines of text into a LaTeX list.
}
\details{
This function formats text for use with LaTeX as a list.
}
\section{Warning}{
 Ligatures parsing is very good, however, these elements may
be incorrect.  If a warning is thrown check the use of "ff", "fi", "fl",
"ffi" and "ffl".
}
\examples{
x <- c(
"one, two buckle my shoe",
"three, four close the door",
"five, six pick up sticks")
LL(, x)
LL(FALSE, x)
}

