% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/new_report.R
\name{new_report}
\alias{new_report}
\title{Report Template}
\usage{
new_report(report = "report", template = getOption("temp.reports"),
  bib.loc = getOption("bib.loc"), name = getOption("name.reports"),
  github.user = getOption("github.user"),
  sources = getOption("sources.reports"), path = getwd(), AN.xlsx = TRUE,
  present = getOption("present.template"), open = is.global(2),
  github = FALSE, ...)
}
\arguments{
\item{report}{A character vector of length two or one: (1) the main directory
name and (2) sub directory names (i.e., all the file contents will be
imprinted with this name). If the length of \code{report} is one this name
will be used as the main directory name and all sub directories and files.}

\item{template}{A character string of the internal reports template or an
external path to a template in the reports package style.  This argument
allows the user to change the contents of the report directory that is
generated. See \code{templates} for more.}

\item{bib.loc}{Optional path to a \bold{.bib} resource.}

\item{name}{A character string of the user's name to be used on the report.}

\item{github.user}{GitHub user name (character string).}

\item{sources}{A vector of path(s) to other scripts to be sourced in the
report project upon startup (adds this location to the report project's
\bold{.Rprofile}).}

\item{path}{The path to where the project should be created.  Default is the
current working directory.}

\item{AN.xlsx}{logical.  If \code{TRUE} the article notes (AN) and websites
documents will be in .xlsx format.  If \code{FALSE} the document will be a
.csv file.}

\item{present}{The template to be used in the \bold{PRESENTATION}
.Rmd/.Rpres.  This can be one of the types from \code{slidify_templates} ,
\code{"rstudio"}  (this generates a .Rpres file), or a path to an .Rmd/.Rpres
file.  This argument will be overrode if a custom reports template is
supplied with an .Rmd file in the \bold{inst} directory named slidify.Rmd
(\file{~inst/slidify.Rmd}). Or an .Rpres file in the \bold{inst} directory
named rstudio.Rpres (\file{~inst/slidify.Rmd}).}

\item{open}{logical.  If \code{TRUE} the project will be opened in RStudio.
The default is to test if \code{new_report} is being used in the global
environment, if it is then the project directory will be opened.}

\item{github}{logical.  If \code{TRUE} the repo will be sent to public
\href{https://github.com/}{GitHub} account.}

\item{\ldots}{Other arguments passed to \code{\link[slidify]{author}}.}
}
\value{
Creates a report template.
}
\description{
Generate a report/paper template to increase efficiency.
}
\section{Suggestion}{
 The user may want to set \code{\link[base]{options}} for
\code{bib.loc}, \code{github.user}, \code{name.reports}
\code{sources.reports},\code{present.template} and \code{revealjs.loc} in
the user's primary \code{.Rprofile}:
\enumerate{
  \item{\bold{bib.loc} - The path to the users primary bibliography}
  \item{\bold{name.reports} - The name to use on reports}
  \item{\bold{temp.reports} - The primary template to use to generate reports
  (see \code{template})}
  \item{\bold{github.user} - GitHub user name}
  \item{\bold{speed.temp} - A speed dial like interface that allows the
  template argument to take a numeric arguement.  Setting this option takes
  the form of:\cr \code{options(speed.temp=list(`1`="wordpress_rmd", `2`="basic_rmd"))}}
  \item{\bold{sources.reports} - Path(s) to additional files/scripts that
  should be included to be sourced in the project startup}
  \item{\bold{present.template} - Path to, or defualt, .Rmd/.Rpres file template for
  use in as the .Rmd/.Rpres used in the PRESENTATION directory (see
  \code{slidify_templates} for possible non-path arguments)}
}
}

\section{Additional Guide}{
 Introductory video
\url{http://www.youtube.com/watch?v=ArHQjQyIS70}
}
\examples{
## new_report()

## fx <- folder(delete_me)
## owd <- getwd(); setwd(fx)
## x <- rdirs(admin, 1:15, c("d", "f", "w"), c(1, 4, 6), text.only = TRUE)
## lapply(x, new_report)
## setwd(owd); delete(fx)
}
\references{
\href{https://github.com/ramnathv/slidifyExamples/tree/gh-pages/examples}{slidify examples}  \cr
\href{http://www.rstudio.com/ide/docs/presentations/overview}{RStudio presentations}
}
\seealso{
\code{\link[reports]{doc_temp}},
\code{\link[reports]{presentation}},
\code{\link[reports]{templates}},
\code{\link[reports]{slidify_templates}},
\code{\link[slidify]{author}}

\href{https://github.com/hakimel/reveal.js/}{Installation section of reveal.js GitHub}
}

