\name{new_report}
\alias{new_report}
\title{Report Template}
\usage{
  new_report(report = "report",
    template = getOption("temp.reports"),
    bib.loc = getOption("bib.loc"),
    name = getOption("name.reports"),
    github.user = getOption("github.user"),
    sources = getOption("sources.reports"), path = getwd(),
    AN.xlsx = TRUE, slidify.clean = FALSE, ...)
}
\arguments{
  \item{report}{A character vector of length two or one:
  (1) the main directory name and (2) sub directory names
  (i.e., all the file contents will be imprinted with this
  name). If the length of \code{report} is one this name
  will be used as the main directory name and all sub
  directories and files.}

  \item{template}{A character string of the internal
  reports template or an external path to a template in the
  reports package style.  This argument allows the user to
  change the contents of the report directory that is
  generated. See \code{templates} for more.}

  \item{bib.loc}{Optional path to a .bib resource.}

  \item{path}{The path to where the project should be
  created.  Default is the current working directory.}

  \item{name}{A character string of the user's name to be
  used on the report.}

  \item{github.user}{GitHub user name (character string).}

  \item{sources}{A vector of path(s) to other scripts to be
  sourced in the report project upon startup (adds this
  location to the report project's \code{.Rprofile}).}

  \item{AN.xlsx}{logical.  If \code{TRUE} the article notes
  (AN) will be in .xlsx format.  If \code{FALSE} the
  document will be a .csv file.}

  \item{solidify.clean}{logical.  If \code{TRUE} the
  PRESENTATION .Rmd file will be generated by slidify.  If
  \code{FALSE} the .Rmd file will be initialize with
  additional parameters for convenience.}

  \item{\ldots}{Other arguments passed to
  \code{\link[slidify]{author}}.}
}
\value{
  Creates a report template.
}
\description{
  Generate a report/paper template to increase efficiency.
}
\section{Suggestion}{
  The user may want to set \code{\link[base]{options}} for
  \code{bib.loc}, \code{name.reports}, \code{temp.reports},
  \code{github.user} and \code{sources.reports} in the
  user's primary \code{.Rprofile}: \enumerate{
  \item{\bold{bib.loc} - The path to the users primary
  bibliography} \item{\bold{name.reports} - The name to use
  on reports} \item{\bold{temp.reports} - The primary
  template to use to generate reports (see
  \code{template})} \item{\bold{github.user} - GitHub user
  name} \item{\bold{speed.temp} - A speed dial like
  interface that allows the template argument to ake a
  numeric arguement.  Setting this option takes the form
  of: \code{options(speed.temp=list(`1`="wordpress_rmd",
  `2`="basic_rmd"))}} \item{\bold{sources.reports} -
  Path(s) to additional files/scripts that should be
  included to be sourced in the project startup} }
}

\section{Additional Guide}{
  Introductory video \url{http://youtu.be/cokNUTGtoM4}
}
\examples{
## new_report()
}
\seealso{
  \code{\link[reports]{doc_temp}},
  \code{\link[reports]{presentation}},
  \code{\link[reports]{templates}},
  \code{\link[slidify]{author}}
}

