\name{notes}
\alias{notes}
\title{View Notes}
\usage{
  notes(col.width = 70, notes.col = FALSE)
}
\arguments{
  \item{col.width}{An integer value of the maximum width of
  columns.}

  \item{notes.col}{logical.  If \code{TRUE} the notes
  column will be displayed.}
}
\value{
  Returns a truncated view of user notes.
}
\description{
  Allows a truncated view of your notes
  (main_dir/ARTICLES/notes).
}
\examples{
\dontrun{
notes()
}
}

