\name{notes}
\alias{notes}
\title{View a Notes}
\usage{
  notes(col.width = 70)
}
\arguments{
  \item{col.width}{An integer value of the maximum width of
  columns.}
}
\value{
  Returns a truncated view of user notes.
}
\description{
  Allos a truncated view of your notes
  (main_dir/ARTICLES/notes).
}
\examples{
\dontrun{
notes()
}
}

