% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/QC.R
\name{QC}
\alias{QC}
\title{Quick Convert: Pandoc Convert a String}
\usage{
QC(to = "latex", from = "markdown", text = "clipboard",
  copy2clip = interactive())
}
\arguments{
\item{to}{Pandoc type you're converting to (defualt is \code{"latex"}).}

\item{from}{Pandoc type you're converting from (defualt is \code{"markdown"}).}

\item{text}{A character vector or text copied to the clipboard.  Default is to
read from the clipboard.}

\item{copy2clip}{logical.  If \code{TRUE} attempts to copy the output to the
clipboard.}
}
\value{
Returns a character vector every word capitalized.
}
\description{
Uses Pandoc to convert from one form to another.
}
\details{
This capitalizes every word of a string.
}
\section{Warning}{
 Ligatures parsing is very good, however, these elements may
be incorrect.  If a warning is thrown check the use of "ff", "fi", "fl",
"ffi" and "ffl".
}
\examples{
x <- "\\\\*note: I *like* chocolate **milk** too ***much***!"
QC(text=x)
QC(text=x, "html")
}
\author{
Michael Nelson and Tyler Rinker<tyler.rinker@gmail.com>.
}
\references{
\href{http://johnmacfarlane.net/pandoc/demos.html}{Pandoc Website}
\url{http://stackoverflow.com/a/15539563/1000343}
}

