\docType{package}
\name{reports}
\alias{package-reports}
\alias{reports}
\alias{reports-package}
\title{reports: Assists in the workflow of writing academic articles and other reports}
\description{
  reports is a package that assists in writing reports and
  presentations by providing a frame work that brings
  together existing R, LaTeX/.docx and Pandoc tools.  The
  package is designed to be used with RStudio, MiKTex/Tex
  Live/LibreOffice, knitr, knitcitations, Pandoc and
  pander.  The user will want to download these free
  programs to maximize the effectiveness of the reports
  package.
}

