% Generated by roxygen2 (4.0.0): do not edit by hand
\name{HS}
\alias{HS}
\alias{VS}
\title{Insert HTML Space}
\usage{
HS(n = 1, copy2clip = interactive())

VS(n = 1, copy2clip = interactive())
}
\arguments{
\item{n}{Number of spaces to insert.}

\item{copy2clip}{logical.  If \code{TRUE} attempts to copy the output to the
clipboard.}
}
\description{
Insert n iterations of HTML spacing into a document.
}
\details{
\code{HS} returns horizontal space (using \code{&nbsp;}) and \code{VS}
returns vertical space (using \code{</br>}).
}
\examples{
paste0("reports", HS(10), "end")
cat(paste0("the", VS(), "end"))
cat(paste0("the", VS(3), "end"))
}

