\docType{data}
\name{themes}
\alias{themes}
\title{themes}
\format{A character vector with 28 elements}
\usage{
data(themes)
}
\description{
a vector of available Beamer themes
}
\details{
\href{http://deic.uab.es/~iblanes/beamer_gallery/index_by_theme.html}{http://deic.uab.es/~iblanes/beamer_gallery/index_by_theme.html}
}
\references{
\url{http://deic.uab.es/~iblanes/beamer_gallery/index_by_theme.htm}
}
\keyword{datasets}

